backend/requirements.txt
Django

djangorestframework
markdown
django-filter
drf-nested-routers
djangorestframework-simplejwt

pyyaml
uritemplate

googlemaps
requests

mysqlclient
redis

gunicorn
uvicorn